+++ changelog.rst
@@ -1,4 +1,13 @@
-version 0.2.0 (2014-08-16)
+version 0.3.0 (2017-03-22)
+================
+
+* Changed the behavior of methods as a workaround for PHP SSH2 bug:
+https://bugs.php.net/bug.php?id=71376
+Restored previous behavior with `intval` of SFTP Resource Handle.
+
+* Merged PR #6 by `pedrofornaza`
+
+version 0.2.0 (2017-02-09)
 ================
 
 * Changed the behavior of methods as a workaround for PHP SSH2 bug: